dev_tools/appsec-mcp/requirements.txt
mcp[cli]>=1.0.0
cerebras-cloud-sdk>=1.0.0
